order/requirements.txt
Flask==2.1.1
redis==4.2.2
gunicorn==20.1.0
kafka-python==2.0.2